+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* `08_error_handling_basics` module complete - exercises, mirrored solutions and exercise hints for `error1`/`error2`
 * `07_strings` module complete - exercises, mirrored solutions and exercise hints for `string1`/`string2`.
 * `06_scope` module complete - exercises, mirrored solutions and exercise hints for `scope1`/`scope2`.
 * `05_tables_basics` module complete - exercises, mirrored solutions and exercise hints for `tables1`-`tables3`.